+++ src/components/Container.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="mx-a w-full p-sm lg:w-4xl md:w-2xl sm:w-xl">
+    <slot name="default" />
+  </div>
+</template>